backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
SQLAlchemy
httpx
python-dotenv
pyodbc